backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic>=2.8.0
python-multipart==0.0.6
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
supabase==2.27.2
# 注：Gemini 模型通过 OpenRouter API 调用，无需额外 SDK

